+++ haoskiosk/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## v1.0.1 - August 2025
+
+- Simplified and generalzedlibinput discovery tagging and merged resulting
+  code into 'run.sh' (Thanks to GuntherSchulz01 and tacher4000)
+- Added "CURSOR_TIMEOUT" to hide cursor (Thanks tacher4000)
+- Set LANG consistent with keyboard layout (Thanks tacher4000)
+- Added additional logging to help debug any future screen or input (touch
+  or mouse) issues
+- Substituted luakit browser-level Dark Mode preference for HA-specific
+  theme preference (Thanks tacher4000)
+
+## v1.0.0 - July 2025
+
+- Switched from (legacy) framebuffer-based video (fbdev) to OpenGL/DRI
+  video
+- Switched from (legacy) evdev input handling to libinput input handling
+- Switched from "HDMI PORT" to "OUTPUT NUMBER" to determine which physical
+  port is displayed
+- Added 'rotation' config to rotate display
+- Added boolean config to determine whether touch inputs are mapped to the
+  display output (in particular, this will rotate them in sync)
+- Modified 'xorg.conf' for consistency with 'OpenGL/DRI' and 'libinput'
+- Attempted to maximize compatibility across RPi and x86
+- Added ability to append to or replace default 'xorg.conf'
+- Added ability to set keyboard layout. (default: 'us')
+- Updated & improved userconf.lua code
+- Extensive changes and improvements to 'run.sh' code
+- Added back (local) DBUS to allow for inter-process luakit communication
+  (e.g., to allow use of unique instance)
+
 ## v0.9.9 - July 2025
 
 - Removed remounting of /dev/ ro (which caused HAOS updates to fail)